added tag picker class, menus and two new screens
