+++ app/src/main/res/layout/edit_tag.xml
@@ -1,84 +1,93 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginHorizontal="10dp"
-    android:orientation="vertical"
-    android:paddingHorizontal="30dp">
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/tagNameInputLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:hint="Tag Name"
-        app:boxCornerRadiusBottomEnd="50dp"
-        app:boxCornerRadiusBottomStart="50dp"
-        app:boxCornerRadiusTopEnd="50dp"
-        app:boxCornerRadiusTopStart="50dp"
-        app:endIconMode="clear_text"
-        app:startIconContentDescription="Notes">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/tagNameInput"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ellipsize="start"
-            android:gravity="center"
-            android:importantForAutofill="no"
-            android:inputType="text"
-            android:maxLength="15"
-            android:paddingVertical="10dp"
-            android:textSize="24sp" />
-
-        <ImageView
-            android:id="@+id/tagColor"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:translationX="15dp"
-            android:translationY="-38dp"
-            app:srcCompat="@drawable/ic_baseline_circle_24" />
-
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/tagColorButton"
-        style="@style/Widget.Material3.Button.TextButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="start"
-        android:layout_marginHorizontal="6dp"
-        android:layout_marginTop="-20dp"
-        android:text="Tag color"
-        app:icon="@drawable/ic_baseline_color_24" />
-
-    <com.google.android.material.chip.ChipGroup
-        android:id="@+id/tagCollection"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        app:selectionRequired="true"
-        app:singleSelection="true">
-
-        <TextView
-            android:id="@+id/tapBlurb"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingHorizontal="5dp"
-            android:text="Recent tags"
-            android:textStyle="normal"
-            android:visibility="visible" />
-
-    </com.google.android.material.chip.ChipGroup>
+    android:layout_marginHorizontal="15dp"
+    android:orientation="vertical">
 
     <LinearLayout
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginTop="15dp"
+        android:orientation="vertical">
+
+        <com.google.android.material.card.MaterialCardView
+            android:id="@+id/LoginCard"
+
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+
+            android:background="?android:attr/selectableItemBackground"
+            android:orientation="vertical"
+            app:cardCornerRadius="15dp"
+            app:cardElevation="3dp"
+
+            app:cardPreventCornerOverlap="true"
+            app:cardUseCompatPadding="true"
+            app:strokeColor="@color/transparent">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:orientation="horizontal">
+
+                <ImageView
+                    android:id="@+id/SiteIcon"
+                    android:layout_width="50dp"
+                    android:layout_height="50dp"
+                    android:layout_gravity="center"
+                    android:padding="15dp"
+                    app:srcCompat="@drawable/ic_baseline_circle_24"
+                    app:tint="?android:attr/textColorPrimary" />
+
+                <EditText
+                    android:id="@+id/editTextTextEmailAddress"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:layout_gravity="center_vertical"
+                    android:layout_marginEnd="15dp"
+                    android:ems="10"
+                    android:hint="Enter tag name"
+                    android:inputType="textEmailAddress"
+                    android:maxLines="1"
+                    android:singleLine="true" />
+
+            </LinearLayout>
+
+        </com.google.android.material.card.MaterialCardView>
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="end"
         android:layout_marginVertical="15dp"
         android:orientation="horizontal">
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="10dp"
+            android:layout_weight="1"
+            android:orientation="horizontal">
+
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_gravity="center"
+                app:tint="?attr/colorAccent"
+                app:srcCompat="@drawable/ic_baseline_palette_24" />
+
+            <com.google.android.material.button.MaterialButton
+                android:id="@+id/addColorButton"
+                style="@style/Widget.Material3.Button.TextButton"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Add color" />
+        </LinearLayout>
 
         <com.google.android.material.button.MaterialButton
             android:id="@+id/backButton"
@@ -86,17 +95,15 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="start"
-            android:layout_weight="1"
             android:text="Go back" />
 
         <com.google.android.material.button.MaterialButton
-            android:id="@+id/addTagButton"
+            android:id="@+id/saveTagButton"
             style="@style/Widget.Material3.Button.TextButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="start"
-            android:layout_weight="1"
-            android:text="Save tag" />
+            android:text="Save" />
 
     </LinearLayout>
